+++ CHANGELOG.md
@@ -143,6 +143,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* Update `software-upgrade-without-kernel` for `yay` or `pacman` by providing a generic, comma separated list of packages to ignore
 * Add starting of reflector.service if last system update was to old
 * Extended `bin/configure_local_setting.sh`
 * Added optional parameters "number of retries if host is down" and "number of seconds between retries" for function `net_bazzline_network_ssh_to_host`